Comment out lines not needed
Comment out lines where the program saves arrays and does
operations we do not need in order to save both memory and
computational time.
Keep only the code related to eplus and eminus observables.
Do not use the sparse option because it is now useless both
(given we are already stripping the dataframe from the variables
we do not need) and would raise an error (we did not deal with
all the code, but only with the main operations encountered).